% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/summary.R
\name{print.summary.mf_model}
\alias{print.summary.mf_model}
\title{Print a Mixed-Frequency Model Summary}
\usage{
\method{print}{summary.mf_model}(x, ...)
}
\arguments{
\item{x}{A \code{"summary.mf_model"} object returned by \code{\link[=summary.mf_model]{summary.mf_model()}}.}

\item{...}{Unused.}
}
\value{
\code{x}, invisibly.
}
\description{
Print a Mixed-Frequency Model Summary
}
\examples{
gdp_growth <- tsbox::ts_pc(gdp)
gdp_growth <- tsbox::ts_na_omit(gdp_growth)
model <- mf_model(
  target = gdp_growth,
  indic = baro,
  indic_predict = "auto.arima",
  indic_aggregators = "mean",
  h = 1
)

print(summary(model))
}
